Move with code interface and trait into State namespace
