+++ src/Sylius/Bundle/CoreBundle/DataFixtures/Factory/LocaleFactoryInterface.php
@@ -13,6 +13,7 @@ declare(strict_types=1);
 
 namespace Sylius\Bundle\CoreBundle\DataFixtures\Factory;
 
+use Sylius\Bundle\CoreBundle\DataFixtures\Factory\State\WithCodeInterface;
 use Sylius\Component\Locale\Model\LocaleInterface;
 use Zenstruck\Foundry\ModelFactory;
 use Zenstruck\Foundry\Proxy;
