+++ src/Sylius/Bundle/CoreBundle/DataFixtures/Factory/State/WithCodeTrait.php
@@ -11,7 +11,7 @@
 
 declare(strict_types=1);
 
-namespace Sylius\Bundle\CoreBundle\DataFixtures\Factory;
+namespace Sylius\Bundle\CoreBundle\DataFixtures\Factory\State;
 
 use Zenstruck\Foundry\ModelFactory;
